refactor(llm): split processor responsibilities
